safari-600.7-branch/LayoutTests/fast/storage/storage-detached-iframe-expected.txt

CONSOLE MESSAGE: line 25: Expected exception caught.
CONSOLE MESSAGE: line 39: Expected exception caught.
Bug: https://bugs.webkit.org/show_bug.cgi?id=57140

Description: Crash from null pointer dereference below WebCore::StorageAreaImpl::setItem()

Expect: passed

Result: passed
